Added car image upload, damage request, and rent request api

+++ Backend/Car_Rental_System/Models/DamageCarFine.cs
@@ -4,7 +4,7 @@
     {
         public string Damage_Id { get; set; }
         public string Car_Id { get; set; }
-        public decimal Fine_Amount { get; set; }
+        public double Fine_Amount { get; set; }
 
     }
 }
